hustle/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
python-jose[cryptography]
python-multipart
pydantic-settings
websockets
requests
aiohttp
python-dotenv
